add api activity api add Activity

+++ entity/model/wallet.go
@@ -2,9 +2,9 @@ package model
 
 type Wallet struct {
 	UserId   string `json:"UserId,omitempty" validate:"required"`
-	WalletId int    `json:"WalletId,omitempty"`
+	WalletId uint   `json:"WalletId,omitempty"`
 	Name     string `json:"Name,omitempty" validate:"required,max=25,min=3"`
 	Type     string `json:"Type,omitempty" validate:"required,max=3,min=3"`
 	IsActive string `json:"-"`
-	Amount   int8   `json:"Amount" validate:"required,numeric,gte=0"`
+	Amount   uint32 `json:"Amount" validate:"required,numeric,gte=0"`
 }
